Module rplidar_zenoh_driver::foxglove

protobuf

Modules

line_primitive

Nested message and enum types in LinePrimitive.

location_fix

Nested message and enum types in LocationFix.

log

Nested message and enum types in Log.

packed_element_field

Nested message and enum types in PackedElementField.

points_annotation

Nested message and enum types in PointsAnnotation.

scene_entity_deletion

Nested message and enum types in SceneEntityDeletion.

Structs

ArrowPrimitive

A primitive representing an arrow

CameraCalibration

Camera calibration parameters

CircleAnnotation

A circle annotation on a 2D image

Color

A color in RGBA format

CompressedImage

A compressed image

CubePrimitive

A primitive representing a cube or rectangular prism

CylinderPrimitive

A primitive representing a cylinder, elliptic cylinder, or truncated cone

FrameTransform

A transform between two reference frames in 3D space

FrameTransforms

An array of FrameTransform messages

GeoJson

GeoJSON data for annotating maps

Grid

A 2D grid of data

ImageAnnotations

Array of annotations for a 2D image

KeyValuePair

A key with its associated value

LaserScan

A single scan from a planar laser range-finder

LinePrimitive

A primitive representing a series of points connected by lines

LocationFix

A navigation satellite fix for any Global Navigation Satellite System

Log

A log message

ModelPrimitive

A primitive representing a 3D model file loaded from an external URL or embedded data

PackedElementField

A field present within each element in a byte array of packed elements.

Point2

A point representing a position in 2D space

Point3

A point representing a position in 3D space

PointCloud

A collection of N-dimensional points, which may contain additional fields with information like normals, intensity, etc.

PointsAnnotation

An array of points on a 2D image

Pose

A position and orientation for an object or reference frame in 3D space

PoseInFrame

A timestamped pose for an object or reference frame in 3D space

PosesInFrame

An array of timestamped poses for an object or reference frame in 3D space

Quaternion

A quaternion representing a rotation in 3D space

RawImage

A raw image

SceneEntity

A visual element in a 3D scene. An entity may be composed of multiple primitives which all share the same frame of reference.

SceneEntityDeletion

Command to remove previously published entities

SceneUpdate

An update to the entities displayed in a 3D scene

SpherePrimitive

A primitive representing a sphere or ellipsoid

TextAnnotation

A text label on a 2D image

TextPrimitive

A primitive representing a text label

TriangleListPrimitive

A primitive representing a set of triangles or a surface tiled by triangles

Vector2

A vector in 2D space that represents a direction only

Vector3

A vector in 3D space that represents a direction only